--- a/torchft/checkpointing/pg_transport_test.py
+++ b/torchft/checkpointing/pg_transport_test.py
@@ -1,5 +1,4 @@
 from datetime import timedelta
-from typing import Dict
 from unittest import TestCase, skipUnless
 
 import torch
@@ -7,7 +6,10 @@ from torch.distributed import TCPStore
 
 from torchft.checkpointing.pg_transport import PGTransport
 from torchft.checkpointing.transport import CheckpointTransport
-from torchft.checkpointing.transport_test import run_multi_recovery_test
+from torchft.checkpointing.transport_test import (
+    make_state_dict,
+    run_multi_recovery_test,
+)
 from torchft.process_group import ProcessGroupBabyNCCL, ProcessGroupGloo
 
 
@@ -18,7 +20,7 @@ class PGTransportTest(TestCase):
         )
         device: torch.device = torch.device("cpu")
 
-        def init(rank: int, world_size: int) -> CheckpointTransport[Dict[str, object]]:
+        def init(rank: int, world_size: int) -> CheckpointTransport[dict[str, object]]:
             pg = ProcessGroupGloo()
             pg.configure(
                 store_addr=f"localhost:{store.port}/prefix",
@@ -26,7 +28,7 @@ class PGTransportTest(TestCase):
                 world_size=world_size,
             )
 
-            return PGTransport[Dict[str, object]](
+            return PGTransport[dict[str, object]](
                 pg, timeout=timedelta(seconds=10), device=device
             )
 
@@ -39,19 +41,49 @@ class PGTransportTest(TestCase):
             host_name="localhost", port=0, is_master=True, wait_for_workers=False
         )
         device: torch.device = torch.device("cuda")
+        timeout: timedelta = timedelta(seconds=10)
 
-        def init(rank: int, world_size: int) -> CheckpointTransport[Dict[str, object]]:
+        def init(rank: int, world_size: int) -> CheckpointTransport[dict[str, object]]:
             torch.cuda.set_device(rank)
 
-            pg = ProcessGroupBabyNCCL()
+            pg = ProcessGroupBabyNCCL(timeout=timeout)
             pg.configure(
                 store_addr=f"localhost:{store.port}/prefix",
                 rank=rank,
                 world_size=world_size,
             )
 
-            return PGTransport[Dict[str, object]](
-                pg, timeout=timedelta(seconds=10), device=device
+            return PGTransport[dict[str, object]](pg, timeout=timeout, device=device)
+
+        run_multi_recovery_test(self, init, device=device)
+
+    # pyre-fixme[56]: Pyre was not able to infer the type of argument
+    @skipUnless(torch.cuda.device_count() >= 3, "need three CUDA devices")
+    def test_pg_transport_baby_nccl_inplace(self) -> None:
+        store: TCPStore = TCPStore(
+            host_name="localhost", port=0, is_master=True, wait_for_workers=False
+        )
+        device: torch.device = torch.device("cuda")
+        timeout: timedelta = timedelta(seconds=10)
+
+        def state_dict() -> dict[str, object]:
+            return make_state_dict(device)
+
+        def init(rank: int, world_size: int) -> CheckpointTransport[dict[str, object]]:
+            torch.cuda.set_device(rank)
+
+            pg = ProcessGroupBabyNCCL(timeout=timeout)
+            pg.configure(
+                store_addr=f"localhost:{store.port}/prefix",
+                rank=rank,
+                world_size=world_size,
+            )
+
+            return PGTransport[dict[str, object]](
+                pg,
+                timeout=timeout,
+                device=device,
+                state_dict=state_dict,
             )
 
         run_multi_recovery_test(self, init, device=device)