This adds a new `state_dict` argument to `PGTransport` that when provided will give a state_dict to use for doing in-place tensor operations. This has been measured at ~3x faster.


# Test plan:

Correctness:

```
pytest torchft/checkpointing/pg_transport_test.py
```

The improvements have been measured via `pg_transport_bench`.

For inplace operation we see it at ~3x faster (15s -> 4-5s) for 12GB with 3MB tensors size. The remaining overhead is primarily from torchft ProcessGroupBaby queue communication and not proportional to the size of the tensors. 

Reducing this overhead requires some careful consideration and will be addressed in a follow up PR.

```
python torchft/checkpointing/pg_transport_bench.py --device cuda
python torchft/checkpointing/pg_transport_bench.py --device cuda --inplace
```

## inplace

### 12GB/3MB (4k tensors)

```
INFO:torchft.checkpointing.pg_transport:send_checkpoint took 5.05398303642869s
INFO:torchft.checkpointing.pg_transport:recv_checkpoint took 5.637796577066183s
```

### 16KB/4B (4k tensors)

```
INFO:torchft.checkpointing.pg_transport:send_checkpoint took 4.909562937915325s
INFO:torchft.checkpointing.pg_transport:recv_checkpoint took 4.766054484993219s
```

### 48GB/12MB (4k tensors)

```
INFO:torchft.checkpointing.pg_transport:send_checkpoint took 18.53099210932851s
INFO:torchft.checkpointing.pg_transport:recv_checkpoint took 18.847779247909784s
```

## not inplace

### 12GB/3MB (4k tensors)

```
INFO:torchft.checkpointing.pg_transport:send_checkpoint took 15.791493758559227s
INFO:torchft.checkpointing.pg_transport:recv_checkpoint took 17.16875096037984s
```